+++ vite.playground.config.js
@@ -31,6 +31,7 @@ export default defineConfig({
         resolve(import.meta.dirname, 'playground/components/dialog.html'),
         resolve(import.meta.dirname, 'playground/components/popover-menu.html'),
         resolve(import.meta.dirname, 'playground/components/tooltip.html'),
+        resolve(import.meta.dirname, 'playground/components/toast.html'),
       ],
     },
   },